Création UserGroupService et liste des groupes dans la page territoire
